requirements.txt
langchain
langchain-community
langchain-aws
boto3
faiss-cpu
pymupdf
python-dotenv
